Extract BigQuery view materialization cache to separate class
