+++ plugin/trino-bigquery/src/main/java/io/trino/plugin/bigquery/BigQueryClient.java
@@ -13,7 +13,6 @@
  */
 package io.trino.plugin.bigquery;
 
-import com.google.cloud.BaseServiceException;
 import com.google.cloud.bigquery.BigQuery;
 import com.google.cloud.bigquery.BigQueryException;
 import com.google.cloud.bigquery.Dataset;
@@ -35,7 +34,6 @@ import com.google.common.collect.ImmutableSet;
 import io.airlift.log.Logger;
 import io.airlift.units.Duration;
 import io.trino.spi.TrinoException;
-import io.trino.spi.connector.SchemaTableName;
 import io.trino.spi.connector.TableNotFoundException;
 
 import java.util.Collections;
@@ -44,8 +42,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutionException;
 import java.util.function.Supplier;
 
 import static com.google.cloud.bigquery.TableDefinition.Type.TABLE;
@@ -55,12 +51,9 @@ import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.Iterables.getOnlyElement;
 import static com.google.common.collect.Streams.stream;
 import static io.trino.plugin.bigquery.BigQueryErrorCode.BIGQUERY_AMBIGUOUS_OBJECT_NAME;
-import static io.trino.plugin.bigquery.BigQueryErrorCode.BIGQUERY_VIEW_DESTINATION_TABLE_CREATION_FAILED;
-import static io.trino.plugin.bigquery.BigQueryUtil.convertToBigQueryException;
 import static java.lang.String.format;
 import static java.util.Locale.ENGLISH;
 import static java.util.Objects.requireNonNull;
-import static java.util.UUID.randomUUID;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.stream.Collectors.joining;
 
@@ -69,18 +62,16 @@ public class BigQueryClient
     private static final Logger log = Logger.get(BigQueryClient.class);
 
     private final BigQuery bigQuery;
-    private final Optional<String> viewMaterializationProject;
-    private final Optional<String> viewMaterializationDataset;
+    private final ViewMaterializationCache materializationCache;
     private final boolean caseInsensitiveNameMatching;
     private final Cache<String, Optional<RemoteDatabaseObject>> remoteDatasets;
     private final Cache<TableId, Optional<RemoteDatabaseObject>> remoteTables;
-    private final Cache<String, TableInfo> destinationTableCache;
 
-    public BigQueryClient(BigQuery bigQuery, BigQueryConfig config)
+    public BigQueryClient(BigQuery bigQuery, BigQueryConfig config, ViewMaterializationCache materializationCache)
     {
         this.bigQuery = requireNonNull(bigQuery, "bigQuery is null");
-        this.viewMaterializationProject = requireNonNull(config, "config is null").getViewMaterializationProject();
-        this.viewMaterializationDataset = config.getViewMaterializationDataset();
+        this.materializationCache = requireNonNull(materializationCache, "materializationCache is null");
+
         Duration caseInsensitiveNameMatchingCacheTtl = requireNonNull(config.getCaseInsensitiveNameMatchingCacheTtl(), "caseInsensitiveNameMatchingCacheTtl is null");
 
         this.caseInsensitiveNameMatching = config.isCaseInsensitiveNameMatching();
@@ -88,10 +79,6 @@ public class BigQueryClient
                 .expireAfterWrite(caseInsensitiveNameMatchingCacheTtl.toMillis(), MILLISECONDS);
         this.remoteDatasets = remoteNamesCacheBuilder.build();
         this.remoteTables = remoteNamesCacheBuilder.build();
-        this.destinationTableCache = CacheBuilder.newBuilder()
-                .expireAfterWrite(config.getViewsCacheTtl().toMillis(), MILLISECONDS)
-                .maximumSize(1000)
-                .build();
     }
 
     public Optional<RemoteDatabaseObject> toRemoteDataset(String projectId, String datasetName)
@@ -192,13 +179,7 @@ public class BigQueryClient
     {
         String query = selectSql(remoteTableId, requiredColumns);
         log.debug("query is %s", query);
-        try {
-            return destinationTableCache.get(query,
-                    new DestinationTableBuilder(this, viewExpiration, query, remoteTableId.getTableId()));
-        }
-        catch (ExecutionException e) {
-            throw new TrinoException(BIGQUERY_VIEW_DESTINATION_TABLE_CREATION_FAILED, "Error creating destination table", e);
-        }
+        return materializationCache.getCachedTable(this, query, viewExpiration, remoteTableId);
     }
 
     public String getProjectId()
@@ -220,16 +201,7 @@ public class BigQueryClient
                 .collect(toImmutableList());
     }
 
-    private TableId createDestinationTable(TableId remoteTableId)
-    {
-        String project = viewMaterializationProject.orElse(remoteTableId.getProject());
-        String dataset = viewMaterializationDataset.orElse(remoteTableId.getDataset());
-
-        String name = format("_pbc_%s", randomUUID().toString().toLowerCase(ENGLISH).replace("-", ""));
-        return TableId.of(project, dataset, name);
-    }
-
-    private Table update(TableInfo table)
+    Table update(TableInfo table)
     {
         return bigQuery.update(table);
     }
@@ -254,7 +226,7 @@ public class BigQueryClient
         bigQuery.delete(tableId);
     }
 
-    private Job create(JobInfo jobInfo)
+    Job create(JobInfo jobInfo)
     {
         return bigQuery.create(jobInfo);
     }
@@ -349,65 +321,6 @@ public class BigQueryClient
         public boolean isAmbiguous()
         {
             return remoteNames.size() > 1;
-        }
-    }
-
-    private static class DestinationTableBuilder
-            implements Callable<TableInfo>
-    {
-        private final BigQueryClient bigQueryClient;
-        private final Duration viewExpiration;
-        private final String query;
-        private final TableId remoteTable;
-
-        DestinationTableBuilder(BigQueryClient bigQueryClient, Duration viewExpiration, String query, TableId remoteTable)
-        {
-            this.bigQueryClient = requireNonNull(bigQueryClient, "bigQueryClient is null");
-            this.viewExpiration = requireNonNull(viewExpiration, "viewExpiration is null");
-            this.query = requireNonNull(query, "query is null");
-            this.remoteTable = requireNonNull(remoteTable, "remoteTable is null");
-        }
-
-        @Override
-        public TableInfo call()
-        {
-            return createTableFromQuery();
-        }
-
-        private TableInfo createTableFromQuery()
-        {
-            TableId destinationTable = bigQueryClient.createDestinationTable(remoteTable);
-            log.debug("destinationTable is %s", destinationTable);
-            JobInfo jobInfo = JobInfo.of(
-                    QueryJobConfiguration
-                            .newBuilder(query)
-                            .setDestinationTable(destinationTable)
-                            .build());
-            log.debug("running query %s", jobInfo);
-            Job job = waitForJob(bigQueryClient.create(jobInfo));
-            log.debug("job has finished. %s", job);
-            if (job.getStatus().getError() != null) {
-                throw convertToBigQueryException(job.getStatus().getError());
-            }
-            // add expiration time to the table
-            TableInfo createdTable = bigQueryClient.getTable(destinationTable)
-                    .orElseThrow(() -> new TableNotFoundException(new SchemaTableName(destinationTable.getDataset(), destinationTable.getTable())));
-            long expirationTimeMillis = createdTable.getCreationTime() + viewExpiration.toMillis();
-            Table updatedTable = bigQueryClient.update(createdTable.toBuilder()
-                    .setExpirationTime(expirationTimeMillis)
-                    .build());
-            return updatedTable;
-        }
-
-        private Job waitForJob(Job job)
-        {
-            try {
-                return job.waitFor();
-            }
-            catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new BigQueryException(BaseServiceException.UNKNOWN_CODE, format("Job %s has been interrupted", job.getJobId()), e);
-            }
         }
     }
 }
