+++ plugin/trino-bigquery/src/main/java/io/trino/plugin/bigquery/BigQueryConnectorModule.java
@@ -59,6 +59,7 @@ public class BigQueryConnectorModule
         binder.bind(BigQueryMetadata.class).in(Scopes.SINGLETON);
         binder.bind(BigQuerySplitManager.class).in(Scopes.SINGLETON);
         binder.bind(BigQueryPageSourceProvider.class).in(Scopes.SINGLETON);
+        binder.bind(ViewMaterializationCache.class).in(Scopes.SINGLETON);
         configBinder(binder).bindConfig(BigQueryConfig.class);
     }
 }
